TIP-458: Move readers to folders following the file format
