+++ src/Pim/Bundle/ConnectorBundle/Resources/config/readers.yml
@@ -4,19 +4,19 @@ parameters:
     pim_connector.reader.database.product.class: Pim\Component\Connector\Reader\Database\ProductReader
     pim_connector.reader.database.base.class:    Pim\Component\Connector\Reader\Database\BaseReader
 
-    pim_connector.reader.file.csv.class:               Pim\Component\Connector\Reader\File\CsvReader
-    pim_connector.reader.file.csv_variant_group.class: Pim\Component\Connector\Reader\File\CsvVariantGroupReader
-    pim_connector.reader.file.csv_association.class:   Pim\Component\Connector\Reader\File\CsvAssociationReader
-    pim_connector.reader.file.csv_product.class:       Pim\Component\Connector\Reader\File\Product\CsvProductReader
+    pim_connector.reader.file.csv.class:               Pim\Component\Connector\Reader\File\Csv\CsvReader
+    pim_connector.reader.file.csv_variant_group.class: Pim\Component\Connector\Reader\File\Csv\CsvVariantGroupReader
+    pim_connector.reader.file.csv_association.class:   Pim\Component\Connector\Reader\File\Csv\CsvAssociationReader
+    pim_connector.reader.file.csv_product.class:       Pim\Component\Connector\Reader\File\Csv\CsvProductReader
 
-    pim_connector.reader.file.xlsx.class:         Pim\Component\Connector\Reader\File\XlsxReader
-    pim_connector.reader.file.xlsx_product.class: Pim\Component\Connector\Reader\File\Product\XlsxProductReader
+    pim_connector.reader.file.xlsx.class:         Pim\Component\Connector\Reader\File\Xlsx\XlsxReader
+    pim_connector.reader.file.xlsx_product.class: Pim\Component\Connector\Reader\File\Xlsx\XlsxProductReader
+
+    pim_connector.reader.file.yaml.class: Pim\Component\Connector\Reader\File\Yaml\YamlReader
 
     pim_connector.reader.file.file_iterator.class:          Pim\Component\Connector\Reader\File\FileIterator
     pim_connector.reader.file.file_iterator_factory.class:  Pim\Component\Connector\Reader\File\FileIteratorFactory
     pim_connector.reader.file.media_path_transformer.class: Pim\Component\Connector\Reader\File\MediaPathTransformer
-
-    pim_connector.reader.file.yaml.class: Pim\Component\Connector\Reader\File\YamlReader
 
 services:
     pim_connector.reader.database.product:
