+++ src/Pim/Component/Connector/spec/Reader/File/Csv/CsvProductReaderSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Component\Connector\Reader\File\Product;
+namespace spec\Pim\Component\Connector\Reader\File\Csv;
 
 use Akeneo\Component\Batch\Job\JobParameters;
 use Akeneo\Component\Batch\Model\StepExecution;
@@ -23,12 +23,12 @@ class CsvProductReaderSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('Pim\Component\Connector\Reader\File\Product\CsvProductReader');
+        $this->shouldHaveType('Pim\Component\Connector\Reader\File\Csv\CsvProductReader');
     }
 
     function it_is_a_csv_reader()
     {
-        $this->shouldHaveType('Pim\Component\Connector\Reader\File\CsvReader');
+        $this->shouldHaveType('Pim\Component\Connector\Reader\File\Csv\CsvReader');
     }
 
     function it_transforms_media_paths_to_absolute_paths(
